requirements.txt
flask
mysql-connector-python
google-generativeai
fpdf
python-dotenv
